debug

+++ Agents/DQN.py
@@ -121,7 +121,7 @@ class DQNAgent(torch.nn.Module):
             # Infer
             action = self.actor(self.critic(obs[instruction], creations), self.step).best
 
-            mistake = cross_entropy(action, label[instruction], reduction='none')
+            mistake = cross_entropy(action, label[instruction].long(), reduction='none')
 
             # Supervised loss
             supervised_loss = mistake.mean()
